[Refactor] : 엔티티 Enum 값 매핑 방식 수정
