+++ src/main/java/kr/co/studyhubinu/studyhubserver/apply/enums/Inspection.java
@@ -1,16 +1,16 @@
-package kr.co.studyhubinu.studyhubserver.apply.domain.insepection;
+package kr.co.studyhubinu.studyhubserver.apply.enums;
 
+import kr.co.studyhubinu.studyhubserver.common.enums.EnumModel;
 import kr.co.studyhubinu.studyhubserver.exception.apply.InspectionNotFoundException;
 import lombok.Getter;
 
 import java.util.Arrays;
 
-@Getter
-public enum Inspection {
+public enum Inspection implements EnumModel {
 
-    ACCEPT("Accept"),
-    STANDBY("StandBy"),
-    REJECT("Reject");
+    ACCEPT("수락"),
+    STANDBY("대기"),
+    REJECT("거절");
 
     private String value;
 
@@ -18,10 +18,21 @@ public enum Inspection {
         this.value = value;
     }
 
+
     public static Inspection fromCode(String dbData) {
         return Arrays.stream(Inspection.values())
                 .filter(v -> v.getValue().equals(dbData))
                 .findAny()
                 .orElseThrow(InspectionNotFoundException::new);
+    }
+
+    @Override
+    public String getKey() {
+        return name();
+    }
+
+    @Override
+    public String getValue() {
+        return value;
     }
 }
